requirements.txt
discord==1.7.3
python-dotenv==0.19.0
selenium==3.141.0
undetected_chromedriver==3.0.3
